# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Liquidity Radar Historical MVP",
    "description": "Read-only HISTORICAL data. Forward scores are not calibrated probabilities; exit cost is not observed slippage. Models are never executed by HTTP requests.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/overview": {
      "get": {
        "tags": [
          "Historical risk"
        ],
        "summary": "Overview",
        "operationId": "overview_api_overview_get",
        "parameters": [
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only windows closed at or before this UTC instant.",
              "title": "As Of"
            },
            "description": "Only windows closed at or before this UTC instant."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Overview_"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/market-risk": {
      "get": {
        "tags": [
          "Historical risk"
        ],
        "summary": "Market Risk",
        "operationId": "market_risk_api_market_risk_get",
        "parameters": [
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MarketRisk_"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/forward-risk": {
      "get": {
        "tags": [
          "Historical risk"
        ],
        "summary": "Forward Risk",
        "operationId": "forward_risk_api_forward_risk_get",
        "parameters": [
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ForwardRisk_"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/positions": {
      "get": {
        "tags": [
          "Position exit"
        ],
        "summary": "Positions",
        "operationId": "positions_api_positions_get",
        "parameters": [
          {
            "name": "scenario",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "NORMAL",
                "STRESS_25",
                "STRESS_50",
                "STRESS_75"
              ],
              "type": "string",
              "default": "NORMAL",
              "title": "Scenario"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Positions_"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/exit-cost": {
      "get": {
        "tags": [
          "Position exit"
        ],
        "summary": "Exit Cost",
        "operationId": "exit_cost_api_exit_cost_get",
        "parameters": [
          {
            "name": "position_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "USDC: 10000, 50000, 100000, 250000, 500000, 1000000",
              "default": 100000,
              "title": "Position Size"
            },
            "description": "USDC: 10000, 50000, 100000, 250000, 500000, 1000000"
          },
          {
            "name": "scenario",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "NORMAL",
                "STRESS_25",
                "STRESS_50",
                "STRESS_75"
              ],
              "type": "string",
              "default": "NORMAL",
              "title": "Scenario"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PositionExit_"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/liquidity/history": {
      "get": {
        "tags": [
          "Historical series"
        ],
        "summary": "History",
        "operationId": "history_api_liquidity_history_get",
        "parameters": [
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "1m",
                "5m",
                "15m"
              ],
              "type": "string",
              "default": "1m",
              "title": "Interval"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "minimum": 1,
              "default": 300,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_History_"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/backtest/summary": {
      "get": {
        "tags": [
          "Retrospective evaluation"
        ],
        "summary": "Backtest Summary",
        "operationId": "backtest_summary_api_backtest_summary_get",
        "parameters": [
          {
            "name": "event",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "LOW_VOLUME_EVENT",
                "HIGH_VOLATILITY_EVENT",
                "HIGH_EXIT_COST_EVENT",
                "ANY_DETERIORATION"
              ],
              "type": "string",
              "default": "ANY_DETERIORATION",
              "title": "Event"
            }
          },
          {
            "name": "sample",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "ALL_AVAILABLE",
                "MATCHED",
                "NON_OVERLAPPING_MATCHED",
                "STRICT_COMPLETE_MATCHED"
              ],
              "type": "string",
              "default": "MATCHED",
              "title": "Sample"
            }
          },
          {
            "name": "horizon",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "5, 15, 30 or 60 minutes",
              "title": "Horizon"
            },
            "description": "5, 15, 30 or 60 minutes"
          },
          {
            "name": "predictor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "CURRENT_MARKET_RISK",
                    "FORWARD_LIQUIDITY_RISK",
                    "LOW_CURRENT_VOLUME",
                    "RECENT_VOLUME_DROP",
                    "PERSISTENCE_PREVIOUS_CONDITION",
                    "PREVALENCE_TRAINING"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Predictor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Backtest_"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Envelope_ForwardRisk_": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Metadata"
          },
          "data": {
            "$ref": "#/components/schemas/ForwardRisk"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "title": "Envelope[ForwardRisk]"
      },
...
      "Envelope_MarketRisk_": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Metadata"
          },
          "data": {
            "$ref": "#/components/schemas/MarketRisk"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "title": "Envelope[MarketRisk]"
      },
      "Envelope_Overview_": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Metadata"
          },
          "data": {
            "$ref": "#/components/schemas/Overview"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "title": "Envelope[Overview]"
      },
      "Envelope_PositionExit_": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Metadata"
          },
          "data": {
            "$ref": "#/components/schemas/PositionExit"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "title": "Envelope[PositionExit]"
      },
      "Envelope_Positions_": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Metadata"
          },
          "data": {
            "$ref": "#/components/schemas/Positions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "title": "Envelope[Positions]"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorInfo"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
    }
  }
}